LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/fontface-font-variation-settings-persisted-js-api-expected.txt


FAIL Test FontFace variationSettings property assert_equals: expected (string) "\"wght\" 850" but got (undefined) undefined
